Commit
When we start using jQuery, we will modify HTML elements without
needing to actually change them in HTML.

Let's make sure that everyone knows they shouldn't actually modify any
of this code directly.

Remember that you can start a comment with <!-- and end a comment with
-->

Add a comment at the top of your HTML that says Only change code above
this line.