Commit
We can also change the CSS of an HTML element directly with jQuery.

jQuery has a function called .css() that allows you to change the CSS
of an element.

Here's how we would change its color to blue:

$("#target1").css("color", "blue");

This is slightly different from a normal CSS declaration, because the
CSS property and its value are in quotes, and separated with a comma
instead of a colon.

Delete your jQuery selectors, leaving an empty document ready function.

Select target1 and change its color to red.